+++ crates/ruma-events/src/call.rs
@@ -6,10 +6,9 @@ pub mod answer;
 pub mod candidates;
 pub mod hangup;
 pub mod invite;
-pub mod negotiate;
-
 #[cfg(feature = "unstable-msc3401")]
 pub mod member;
+pub mod negotiate;
 #[cfg(feature = "unstable-msc4075")]
 pub mod notify;
 pub mod reject;
