+++ crates/ruma-events/src/call/member/focus.rs
@@ -1,4 +1,4 @@
-//! Types for matrixRTC Focus/SFU configurations.
+//! Types for MatrixRTC Focus/SFU configurations.
 
 use ruma_macros::StringEnum;
 use serde::{Deserialize, Serialize};
@@ -7,17 +7,17 @@ use crate::PrivOwnedStr;
 
 /// Description of the SFU/Focus a membership can be connected to.
 ///
-/// A focus can be any server powering the matrixRTC session (SFU,
+/// A focus can be any server powering the MatrixRTC session (SFU,
 /// MCU). It serves as a node to redistribute RTC streams.
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 #[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
 #[serde(tag = "type", rename_all = "snake_case")]
 pub enum Focus {
-    /// LiveKit is one possible type of SFU/Focus that can be used for a matrixRTC session.
+    /// LiveKit is one possible type of SFU/Focus that can be used for a MatrixRTC session.
     Livekit(LivekitFocus),
 }
 
-/// The fields to describe LiveKit as a `preferred_foci`.
+/// The struct to describe LiveKit as a `preferred_foci`.
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 #[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
 pub struct LivekitFocus {
@@ -36,7 +36,7 @@ impl LivekitFocus {
     /// # Arguments
     ///
     /// * `alias` - The alias with which the LiveKit sessions can be reached.
-    /// * `service_url` - The url of the jwt server for the LiveKit instance.
+    /// * `service_url` - The url of the JWT server for the LiveKit instance.
     pub fn new(alias: String, service_url: String) -> Self {
         Self { alias, service_url }
     }
@@ -44,13 +44,13 @@ impl LivekitFocus {
 
 /// Data to define the actively used Focus.
 ///
-/// A focus can be any server powering the matrixRTC session (SFU,
+/// A focus can be any server powering the MatrixRTC session (SFU,
 /// MCU). It serves as a node to redistribute RTC streams.
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 #[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
 #[serde(tag = "type", rename_all = "snake_case")]
 pub enum ActiveFocus {
-    /// LiveKit is one possible type of SFU/Focus that can be used for a matrixRTC session.
+    /// LiveKit is one possible type of SFU/Focus that can be used for a MatrixRTC session.
     Livekit(ActiveLivekitFocus),
 }
